+++ src/main/java/com/consultrix/consultrixserver/repository/OrganizationRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface OrganizationRepository extends JpaRepository<Organization, Long> {
+public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
     Optional<Organization> findByNameIgnoreCase(String name);
 }
